% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mermaid_diagram.R
\name{create_mermaid_diagram}
\alias{create_mermaid_diagram}
\title{Create a diagram to show participant progress through a study workflow}
\usage{
create_mermaid_diagram(
  participant_level_progress,
  parents,
  children,
  pretty_labels,
  render = TRUE
)
}
\arguments{
\item{participant_level_progress}{A detail dataframe describing the progress
of each research participant}

\item{parents}{A vector nodes of the diagram that have children}

\item{children}{A vector enumerating the child categorical level below the
corresponding parent}

\item{pretty_labels}{A dataframe defining the pretty labels for each step
and each factor level}

\item{render}{A logical to indicate if the mermaid syntax should be
rendered into a diagram or if the syntax should be returned
(default = TRUE)}
}
\value{
A plot of the study progress or the mermaid syntax to generate it
}
\description{
Create a diagram to show participant progress through a study workflow via
the counts of study participants who reach each point in the study workflow.
Create the diagram diagram from a detailed dataframe, describing the
progress of each research participant, the parent child relationships
that define the study progression, and pretty labels for each state
and step in the protocol.
}
\examples{
\dontrun{
create_mermaid_diagram(
  participant_level_progress = participant_level_progress,
  parents = steps$parent,
  children = steps$child,
  pretty_labels = pretty_labels
)
}
}
