% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{Create a table to summarize participant progress}
\usage{
create_table(participant_level_progress, children, pretty_labels)
}
\arguments{
\item{participant_level_progress}{A detail dataframe describing the progress
of each research participant}

\item{children}{A vector enumerating the child categorical level below the
corresponding parent}

\item{pretty_labels}{A dataframe defining the pretty labels for each step
and each factor level}
}
\value{
a gtsummary::tbl_summary() object that describes
\code{participant_level_progress} using the labels in \code{pretty_labels}
}
\description{
Create a table showing the counts of study participants who reached each point in the study workflow
}
\examples{
\dontrun{
gtsummary_table <- create_table(participant_level_progress, pretty_labels)
}
}
