% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pretty_labels_template.R
\name{get_pretty_labels_template}
\alias{get_pretty_labels_template}
\title{Create a dataframe of steps and factor levels in the study
workflow to facilitate label replacement}
\usage{
get_pretty_labels_template(participant_level_progress, parents, children)
}
\arguments{
\item{participant_level_progress}{A detail dataframe describing the progress
of each research participant}

\item{parents}{A vector nodes of the diagram that have children}

\item{children}{A vector enumerating the child categorical level below the
corresponding parent}
}
\value{
A data frame with on row for each step in the study and one row
for each factor level at each step. The data frame has these columns:
\itemize{
\item variable - the factor for this level
\item row_type - a character string indicating if this row describes a step
of the factor values on that step
\item plain_label - the existing label derived from the step or its factor level
\item pretty_label - the pretty label to use in place of label
}
}
\description{
Generate a dataframe describing the steps and factor levels in the study
workflow to assist in setting pretty labels for each step and factor level.
}
\examples{
\dontrun{
get_pretty_labels_template(
  participant_level_progress = progress_detail,
  parents = steps$parent,
  children = steps$child
)
}
}
