% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_summary_labels.R
\name{get_table_summary_labels}
\alias{get_table_summary_labels}
\title{get_table_summary_labels}
\usage{
get_table_summary_labels(pretty_labels)
}
\arguments{
\item{pretty_labels}{a data like those returned by
\code{\link[=get_pretty_labels_template]{get_pretty_labels_template()}}, but with pretty labels in place of the
templated values.}
}
\value{
a named vector of pretty labels with the plain labels as the names
}
\description{
get a named vector we can use to relabel a gtsummary table
}
\examples{
\dontrun{
  get_table_summary_labels(pretty_labels)
}
}
\keyword{Internal}
