% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_participant_level_progress_pretty.R
\name{get_participant_level_progress_pretty}
\alias{get_participant_level_progress_pretty}
\title{Replace plain text labels with pretty labels in \code{participant_level_progress}}
\usage{
get_participant_level_progress_pretty(
  participant_level_progress,
  pretty_labels
)
}
\arguments{
\item{participant_level_progress}{A detail dataframe describing the progress
of each research participant}

\item{pretty_labels}{a data like those returned by
\code{\link[=get_pretty_labels_template]{get_pretty_labels_template()}}, but with pretty labels in place of the
templated values.}
}
\value{
a dataframe the same structure as \code{participant_level_progress},
but with pretty labels in the data fields in place of the plain
labels.
}
\description{
Replace plain text labels with pretty labels in \code{participant_level_progress}
in preparation for creating a gtsummary table.
}
\examples{
\dontrun{
  get_participant_level_progress_pretty(
    participant_level_progress,
    pretty_labels
  )
}
}
\keyword{Internal}
